els-ASAP auto commit file/tag gstreamer1-plugins-good-1.4.5-2.el7.src.rpm

+++ src/gst/wavparse/gstwavparse.h
@@ -86,7 +86,7 @@ struct _GstWavParse {
   guint16 blockalign;
   guint16 width;
   guint32 av_bps;
-  guint64 fact;
+  guint32 fact;
 
   /* real bps used or 0 when no bitrate is known */
   guint32 bps;
